+++ src/main/java/school/faang/user_service/dto/user/UserDto.java
@@ -1,8 +1,7 @@
-package school.faang.user_service.dto.entity;
+package school.faang.user_service.dto.user;
 
 import lombok.*;
-
-import java.util.List;
+import school.faang.user_service.dto.entity.EntityDto;
 
 @Builder
 @Data
@@ -14,17 +13,9 @@ public class UserDto extends EntityDto {
     private String username;
     private String email;
     private String phone;
-    private String password;
-    private Boolean active;
     private String aboutMe;
     private Long countryId;
     private String city;
     private Integer experience;
-    private List<Long> followersIds;
-    private List<Long> followeesIds;
-    private List<Long> ownedEventsIds;
-    private List<Long> menteesIds;
-    private List<Long> mentorsIds;
-    private List<Long> skillsIds;
 
 }
