+++ src/main/java/school/faang/user_service/filter/memory/user/IdUserInMemoryFilter.java
@@ -1,7 +1,7 @@
 package school.faang.user_service.filter.memory.user;
 
 import org.springframework.stereotype.Component;
-import school.faang.user_service.dto.entity.UserDto;
+import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.dto.filter.UserFilterDto;
 
 import java.util.stream.Stream;
