+++ src/main/java/school/faang/user_service/filter/memory/user/UserInMemoryFilter.java
@@ -1,6 +1,6 @@
 package school.faang.user_service.filter.memory.user;
 
-import school.faang.user_service.dto.entity.UserDto;
+import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.dto.filter.UserFilterDto;
 import school.faang.user_service.filter.memory.InMemoryFilter;
