+++ src/main/java/school/faang/user_service/handler/GlobalExceptionHandler.java
@@ -1,9 +1,10 @@
-package school.faang.user_service.exception;
+package school.faang.user_service.handler;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import school.faang.user_service.exception.DataValidationException;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
