+++ src/main/java/school/faang/user_service/service/mentorship/MentorshipService.java
@@ -2,8 +2,8 @@ package school.faang.user_service.service.mentorship;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import school.faang.user_service.dto.entity.UserDto;
-import school.faang.user_service.mapper.UserMapper;
+import school.faang.user_service.dto.user.UserDto;
+import school.faang.user_service.mapper.user.UserMapper;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.repository.UserRepository;
