+++ src/main/java/school/faang/user_service/service/subscription/SubscriptionService.java
@@ -1,6 +1,6 @@
 package school.faang.user_service.service.subscription;
 
-import school.faang.user_service.dto.entity.UserDto;
+import school.faang.user_service.dto.user.UserDto;
 import school.faang.user_service.dto.filter.UserFilterDto;
 
 import java.util.List;
@@ -15,7 +15,7 @@ public interface SubscriptionService {
 
     int getFollowersCount(long followeeId);
 
-    List<UserDto> getFollowing(long followerId, UserFilterDto filter);
+    List<UserDto> getFollowings(long followerId, UserFilterDto filter);
 
     List<UserDto> getFollowers(long followeeId, UserFilterDto filterDto);
 }
